bot/requirements.txt
python-telegram-bot>=21.0
httpx>=0.25.0
pydantic>=2.0.0
apscheduler>=3.10.0
aiofiles>=23.0.0
tenacity>=8.2.0
async-lru>=2.0.0
